--- broadlink.indigoPlugin/Contents/Server Plugin/broadlink/requirements.txt
@@ -1,1 +1,0 @@
-pycryptodome==3.4.11
